+++ DRN.Framework.Utils/Auth/IScopedUser.cs
@@ -48,6 +48,6 @@ public interface IScopedUser
         if (claimValue is null)
             return defaultValue;
 
-        return claimValue.TryParse<TValue>(out var result) ? result : default;
+        return claimValue.TryParse<TValue>(out var result) ? result : defaultValue;
     }
 }